Adjusted tests
